[WFARQ-74] Switch around the abstracted container configurations so the bootable JAR does not allow the JBoss Home to be set as it's not used.
https://issues.redhat.com/browse/WFARQ-74
